pf tracker to track pages that were called due to page fault and were not UFFD COPY yet

+++ criu/include/pf-tracker.h
@@ -10,8 +10,8 @@ enum pf_state {
 	PF_STATE_COMPLETED,       /* UFFDIO_COPY succeeded */
 };
 
-extern void pf_tracker_add(__u64 address, unsigned long nr_pages, int pid, bool is_pf);
-extern void pf_tracker_set_state(__u64 address, enum pf_state state);
+extern void pf_tracker_add(unsigned long long address, unsigned long nr_pages, int pid, bool is_pf);
+extern void pf_tracker_set_state(unsigned long long address, enum pf_state state);
 extern void pf_tracker_print_stats(void);
 
 #endif /* __CR_PF_TRACKER_H__ */
